Fix threading issues
The chip-8 engine is started from the renderer ARSCNViewDelegate method, which is called on a background thread. This means the chip-8 engine delegate methods are subsequently called on background threads. However, we need to render on the main thread. Rather than moving the whole chip 8 engine onto the main thread we just push setNeedDisplay calls to the main thread.