Reposition in known infinite plane
Repositioning in a calculated plane meant you were trying to position
within the invisible bounds of a currently calculating plane. It worked
fine but not being able to see the bounds wasn't very user friendly.

Solutions were either:
- render bounds
- allow positioning on an infinite plane regardless of bounds

The latter is the easiest thing to do so opting for that