Make screen positioning offset by screen centre
Prior to this the repositioning would jump the centre of the screen to the touch point. Now we reposition relatively so that the touch is offset from the centre of the screen.

We also do away with hit testing. If we're in repositioning mode then the user can touch anywhere on the screen to do this. This is a better user experience as the screen can be quite a small area depending on the distance and angle from the user to the AR screen.