Revert repositioning offset
This was attempting to offset the touch position from the centre of the chip8 screen. However the result seemed relative to how the world origin is instantiated and so we could get issues like the chip8 screen being barely moved even if there were big pan gestures.

Reverting to this means that the screen will always jump its centre to the user's touch for now.